chore(lib/image_resize): 이미지 리사이징 함수 모듈화
